requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytest
pytest-mock
twine
pypandoc
pytest-cov
testfixtures